FIX: unsigned -> int conversion

+++ lib/smurff-cpp/SmurffCpp/DataMatrices/ScarceMatrixData.cpp
@@ -160,7 +160,7 @@ double ScarceMatrixData::sumsq(const SubModel& model) const
    {
       for (Eigen::SparseMatrix<double>::InnerIterator it(Y(), j); it; ++it) 
       {
-         sumsq += square(model.predict({it.row(), it.col()})- it.value());
+         sumsq += square(model.predict({(int)it.row(), (int)it.col()})- it.value());
       }
    }
